requirements.txt
yookassa
aiogram
pillow
apscheduler
aiosqlite
qrcode
aiocryptopay
colorlog
python-dotenv
yoomoney
